// src/api/generated/zod.ts
export const createInflationApiV1InflationPostBodyPeriodYearMin = 1900;
export const createInflationApiV1InflationPostBodyPeriodYearMax = 3000;

export const createInflationApiV1InflationPostBodyPeriodMonthMax = 12;

export const createInflationApiV1InflationPostBodyRateRegExpTwo = new RegExp('^(?!^[-+.]\*$)[+-]?0\*\\d\*\\.?\\d\*$');
export const createInflationApiV1InflationPostBodySourceDefault = "manual";
export const createInflationApiV1InflationPostBodySourceMax = 64;



export const createInflationApiV1InflationPostBody = zod.object({
  "period_year": zod.number().min(createInflationApiV1InflationPostBodyPeriodYearMin).max(createInflationApiV1InflationPostBodyPeriodYearMax),
  "period_month": zod.number().min(1).max(createInflationApiV1InflationPostBodyPeriodMonthMax),
  "rate": zod.union([zod.number(),zod.string().regex(createInflationApiV1InflationPostBodyRateRegExpTwo)]),
  "source": zod.string().max(createInflationApiV1InflationPostBodySourceMax).default(createInflationApiV1InflationPostBodySourceDefault)
})